requirements.txt
streamlit
fastapi
uvicorn
requests
langdetect
sentence-transformers
faiss-cpu
rank-bm25
tqdm
PyMuPDF
python-dotenv